+++ scripts/ci/golden_test.sh
@@ -89,7 +89,8 @@ normalize() {
     -e 's#\[[^]]+/([^/]+\.(cojo|ma|bed|bim|fam|badsnps))\]#[\1]#g' \
     -e 's#/[^ ]*/([^/ ]+\.(cojo|ma|bed|bim|fam|badsnps))#\1#g' \
     -e 's/[0-9]{2}:[0-9]{2}:[0-9]{2}/HH:MM:SS/g' \
-    -e 's/(by GCC )[0-9]+\.[0-9]+/\1X.Y/g'
+    -e 's/(by GCC )[0-9]+\.[0-9]+/\1X.Y/g' \
+    -e 's/(\* version v)[0-9]+\.[0-9]+\.[0-9]+/\1X.Y.Z/g'
 }
 
 fail=0